Hent barn som er registrert på søke fra PDL
