+++ app/utils/formattering.util.ts
@@ -1,0 +1,7 @@
+import { format } from "date-fns";
+import { nb } from "date-fns/locale";
+
+export function formaterNorskDato(dato: Date) {
+  const formattertDato = format(dato, "MM. MMMM yyyy", { locale: nb });
+  return formattertDato.charAt(0).toUpperCase() + formattertDato.slice(1);
+}
